requirements.txt
cognee==0.3.4
python-dotenv>=1.0.0
requests>=2.31.0
